# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Work In Progress",
        "description": "Dokumentasi API untuk proses inspeksi, SWA, dan validasi",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/work-in-progress/by-user": {
            "get": {
                "tags": [
                    "Work In Progress"
                ],
                "summary": "Get Work In Progress list for authenticated user",
                "operationId": "b26be4cfac6744542e089dee2fa9a6a1",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Jumlah item per halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 25
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/work-in-progress/verification/{WIPId}": {
            "post": {
                "tags": [
                    "Work In Progress"
                ],
                "summary": "Verifikasi dan simpan data Stop Work Authority",
                "operationId": "09e49238a423cc26683f988881f3cd1d",
                "parameters": [
                    {
                        "name": "WIPId",
                        "in": "path",
                        "description": "ID dari Work In Progress",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "InputInspection",
                                    "isVerified"
                                ],
                                "properties": {
                                    "InputInspection": {
                                        "type": "object",
                                        "example": {
                                            "field": "value"
                                        }
                                    },
                                    "IsVerified": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Verifikasi berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/work-in-progress/swa/{WIPId}": {
            "get": {
                "tags": [
                    "Work In Progress"
                ],
                "summary": "Lihat daftar SWA berdasarkan WIP",
                "operationId": "f0d2961ecdd7244449bba34a1d8203e1",
                "parameters": [
                    {
                        "name": "WIPId",
                        "in": "path",
                        "description": "ID dari Work In Progress",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "IsWIP",
                        "in": "query",
                        "description": "Filter IsWIP (true/false)",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Jumlah data per halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data SWA berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Work In Progress"
                ],
                "summary": "Create Stop Work Authority (SWA)",
                "operationId": "bfb84f79ffb0881cd86078786dd4c6e8",
                "parameters": [
                    {
                        "name": "WIPId",
                        "in": "path",
                        "description": "ID dari Work In Progress",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "InputInspection (JSON string) dan file upload untuk masing-masing field",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "InputInspection"
                                ],
                                "properties": {
                                    "InputInspection": {
                                        "description": "JSON string dari input inspeksi",
                                        "type": "string",
                                        "example": "[{'key':'sop','answer':'Comply'},{'key':'working_permit','answer':'Comply'}]"
                                    },
                                    "IsDraft": {
                                        "description": "SWA draft atau tidak",
                                        "type": "boolean",
                                        "example": "true"
                                    },
                                    "FileInput[0][key]": {
                                        "description": "Key tag untuk file",
                                        "type": "string",
                                        "example": "sop"
                                    },
                                    "FileInput[0][fileinput]": {
                                        "description": "File PDF untuk key",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "FileInput[1][key]": {
                                        "type": "string",
                                        "example": "working_permit"
                                    },
                                    "FileInput[1][fileinput]": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SWA berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/work-in-progress/swa/{SWAId}/update": {
            "post": {
                "tags": [
                    "Work In Progress"
                ],
                "summary": "Update Stop Work Authority (SWA)",
                "operationId": "451b8b42d49681f9075d90f1e1be4d00",
                "parameters": [
                    {
                        "name": "SWAId",
                        "in": "path",
                        "description": "ID dari SWA yang ingin diupdate",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Form input inspeksi dan file jika ada perubahan",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "InputInspection",
                                    "IsDraft"
                                ],
                                "properties": {
                                    "InputInspection": {
                                        "description": "JSON string dari input inspeksi",
                                        "type": "string",
                                        "example": "[{'key':'sop','answer':'Comply'}]"
                                    },
                                    "IsDraft": {
                                        "description": "Tandai apakah masih draft atau final",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "FileInput[0][key]": {
                                        "type": "string",
                                        "example": "sop"
                                    },
                                    "FileInput[0][fileinput]": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "FileInput[1][key]": {
                                        "type": "string",
                                        "example": "working_permit"
                                    },
                                    "FileInput[1][fileinput]": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SWA berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/work-in-progress/swa/{SWAId}": {
            "delete": {
                "tags": [
                    "Work In Progress"
                ],
                "summary": "Menghapus Stop Work Authority (SWA)",
                "operationId": "595f1de99916dce84a198d03584ea1bd",
                "parameters": [
                    {
                        "name": "SWAId",
                        "in": "path",
                        "description": "ID dari SWA yang akan dihapus",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SWA berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a bearer token to access this API",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}